ci: check for clangd diagnostics
This is a followup from #2340, which adds a check to CI for clangd
diagnostics. It is added to GCC because clang is the slower of the two
linux runners right now.

The script itself spawns clangd and speaks Language Server Protocol to
it. We open files, and clangd returns diagnostics (if any). This enables
a fairly straightforward way of checking an arbitrary number of source
files for warnings/errors.

Simply use `make clangd-diagnostics` to run the tool. The variables
`CLANGD_DIAGNOSTIC_INSTANCES` and `CLANGD_DIAGNOSTIC_JOBS` control the
parallelism.