function: use splitmix64 for hashing
SplitMix64 is an excellent integer hashing function. According to [this
blog][1], it is the main function to beat in terms of hashing. It is
simple and provides much better output than our previous ones.

In particular, this function does a good job of shuffling the higher
bits of the output, a property critical for the new hash index design.