tests: add ability to automatically rewrite test files (#5442)
Include a REWRITE_TESTS flag under the test helper class.

Update the teardown function to rewrite tests if required.

Update the test runner routine to toggle between preforming a diff with the expected results and saving the produced result according to the new flag.

This implementation is not thread safe. As of now, this mode of the runner should only be used with TEST_JOBS=1.

Co-authored-by: CI Bot <tgahunia05@users.noreply.github.com>
Co-authored-by: Siddhartha Sahu <867424+sdht0@users.noreply.github.com>
Co-authored-by: Siddhartha Sahu <dev@sdht.in>
Co-authored-by: CI Bot <sdht0@users.noreply.github.com>